Refactor: use #let, #context, and #describe
context - We've identified that the default tuning and the drop D tuning
are two changes we're making that are common to a group of tests. This
forms a context.
let - We're also repeatedly declaring Guitar.new between these tests.
Using let gives us a block that gets rerun with each test and is memoized.
describe - We're also only calling the #strum method and testing it in
each of these conditions, so we should surround these tests in a #describe
block so that if there's another method on Guitar, we can test it separately.