Use `subject` to fully extract setup outside tests
Since we're describing the `strum` method, it makes sense that the subject
of each test should be a call to the `strum` method. `subject` works identically to `let`.
Unless you pass it a name, you'll find the variable `subject` is now usable in your tests.
`subject` has an added bonus to `let`, though - if you fully commit to having all your setup
outside the body of the test, it lets you declare clean one-liner tests of the form it { is_expected.to ... }.
You can use any RSpec matcher with this.

At this point, we're also using nested contexts, though, which is
advised against. By default, rubocop-rspec will warn you if you're nesting
contexts multiple levels deep, as this can reduce readability.