Extract setup out to shared contexts
To avoid nesting contexts, we can extract out the associated setup. In
this particular case, it might feel redundant at first since we're only
running one test per context, but that could change in future. A better
example to illustrate this would be an ActiveRecord model - you might
want to test that the value of an attribute has changed, whether the
record has been persisted, and whether the record is valid, for example:

it { is_expected.not_to be_persisted }
it { is_expected.not_to be_valid }
it { is_expected.to have_attributes(name: "Daiji Igarashi") }